refactoring and added sanitizer

+++ src/app/app.c
@@ -8,8 +8,6 @@
 #include "../window/window.h"
 #include "../renderer/renderer.h"
 
-#include "../debugmalloc.h"
-
 AppData app_data;
 
 void app_init()
@@ -39,13 +37,13 @@ void app_update()
     app_data.delta_time = (double)(app_data.frame_start - app_data.last_frame_start) / 1000.0;
     app_data.last_frame_start = app_data.frame_start;
 
-    for (int i = 0; i < app_data.windows->size; i++)
+    for (size_t i = 0; i < app_data.windows->size; i++)
         _window_reset((Window*)vector_get(app_data.windows, i));
 
     SDL_Event event;
     while (SDL_PollEvent(&event) != 0)
     {
-        for (int i = 0; i < app_data.windows->size; i++)
+        for (size_t i = 0; i < app_data.windows->size; i++)
         {
             Window* window = (Window*)vector_get(app_data.windows, i);
             if (event.window.windowID == SDL_GetWindowID(window->window))
@@ -55,12 +53,12 @@ void app_update()
             }
         }
     }
-    for (int i = 0; i < app_data.windows->size; i++)
+    for (size_t i = 0; i < app_data.windows->size; i++)
         _window_update((Window*)vector_get(app_data.windows, i));
 }
 void app_render()
 {
-    for (int i = 0; i < app_data.windows->size; i++)
+    for (size_t i = 0; i < app_data.windows->size; i++)
     {
 		Window* window = (Window*)vector_get(app_data.windows, i);
         _window_render(window);
@@ -72,14 +70,14 @@ void app_render()
 }
 void app_request_close()
 {
-    for (int i = 0; i < app_data.windows->size; i++)
+    for (size_t i = 0; i < app_data.windows->size; i++)
         ((Window*)vector_get(app_data.windows, i))->close_requested = true;
 }
 void app_close()
 {
     _font_close();
     _texture_close();
-    for (int i = 0; i < app_data.windows->size; i++)
+    for (size_t i = 0; i < app_data.windows->size; i++)
     {
         Window* window = (Window*)vector_get(app_data.windows, i);
         _window_close(window);
@@ -90,7 +88,7 @@ void app_close()
     TTF_Quit();
     SDL_Quit();
 }
-void app_set_target_fps(int fps)
+void app_set_target_fps(Uint32 fps)
 {
     app_data.target_frame_time = 1000 / fps;
 }
